Settings: Status bar clock and battery enhancements [3/3]
Added to the Clock and Battery Status submenus:
- Show seconds in clock
- Show date in clock, using various styles and formats
- Show clock in various font styles
- Set custom color for clock text
- Set custom colors for battery (fill, charging bolt, percentage text)

Note that previously, when the battery style was set to portrait or landscape
icon, the charging bolt and percentage text (if set to display inside the
battery) were always black. With these changes, the bolt and text will be
black only if the bolt and/or text colors are the same as the fill color;
otherwise, they will use the theme (or custom) colors.