Themes: Partially revert "Themes: Settings: Expose storage icon colors"
LineageOS now exposes the private storage colors, so use those instead.
(The public storage color is still not exposed, so keep that for now.)

The color values for the private storage colors in projekt_colors.xml are
being kept for now, so that themes that change these colors won't break.
Since these colors are no longer used in the code, however, themers will
need to modify the color values from lineage_colors.xml to achieve the
same effect.